Breakup test misc pt2
