+++ test/cpp/jit/test.cpp
@@ -11,10 +11,14 @@
 #include <test/cpp/jit/test_argument_spec.h>
 #include <test/cpp/jit/test_autodiff.h>
 #include <test/cpp/jit/test_class_parser.h>
+#include <test/cpp/jit/test_code_template.h>
 #include <test/cpp/jit/test_constant_pooling.h>
 #include <test/cpp/jit/test_create_autodiff_subgraphs.h>
+#include <test/cpp/jit/test_custom_operators.h>
+#include <test/cpp/jit/test_dynamic_dag.h>
 #include <test/cpp/jit/test_fuser.h>
 #include <test/cpp/jit/test_graph_executor.h>
+#include <test/cpp/jit/test_interpreter.h>
 #include <test/cpp/jit/test_ir.h>
 #include <test/cpp/jit/test_irparser.h>
 #include <test/cpp/jit/test_ivalue.h>
@@ -92,6 +96,5 @@ void runJITCPPTests() {
   testEvalModeForLoadedModule();
 }
 #undef JIT_TEST
-
 } // namespace jit
 } // namespace torch
