+++ test/cpp/jit/test_misc.h
@@ -79,59 +79,6 @@ std::ostream& operator<<(std::ostream& out, const std::vector<T>& list) {
   out << "}";
   return out;
 }
-static const auto ct = CodeTemplate(R"(
-int foo($args) {
-
-    $bar
-        $bar
-    $a+$b
-}
-int commatest(int a${,stuff})
-int notest(int a${,empty,})
-)");
-static const auto ct_expect = R"(
-int foo(hi, 8) {
-
-    what
-    on many
-    lines...
-    7
-        what
-        on many
-        lines...
-        7
-    3+4
-}
-int commatest(int a, things..., others)
-int notest(int a)
-)";
-
-void testCodeTemplate() {
-  {
-    TemplateEnv e;
-    e.s("hi", "foo");
-    e.v("what", {"is", "this"});
-    TemplateEnv c(e);
-    c.s("hi", "foo2");
-    ASSERT_EQ(e.s("hi"), "foo");
-    ASSERT_EQ(c.s("hi"), "foo2");
-    ASSERT_EQ(e.v("what")[0], "is");
-  }
-
-  {
-    TemplateEnv e;
-    e.v("args", {"hi", "8"});
-    e.v("bar", {"what\non many\nlines...", "7"});
-    e.s("a", "3");
-    e.s("b", "4");
-    e.v("stuff", {"things...", "others"});
-    e.v("empty", {});
-    auto s = ct.format(e);
-    // std::cout << "'" << s << "'\n";
-    // std::cout << "'" << ct_expect << "'\n";
-    ASSERT_EQ(s, ct_expect);
-  }
-}
 
 void testInternedStrings() {
   ASSERT_EQ(prim::Param, Symbol::prim("Param"));
@@ -177,30 +124,6 @@ void testFromQualString() {
     } catch (const std::exception& c) {
     }
   }
-}
-
-void testInterp() {
-  constexpr int batch_size = 4;
-  constexpr int input_size = 256;
-  constexpr int seq_len = 32;
-
-  int hidden_size = 2 * input_size;
-
-  auto input = at::randn({seq_len, batch_size, input_size}, at::kCUDA);
-  auto hx = at::randn({batch_size, hidden_size}, at::kCUDA);
-  auto cx = at::randn({batch_size, hidden_size}, at::kCUDA);
-  auto w_ih = t_def(at::randn({4 * hidden_size, input_size}, at::kCUDA));
-  auto w_hh = t_def(at::randn({4 * hidden_size, hidden_size}, at::kCUDA));
-
-  auto lstm_g = build_lstm();
-  Code lstm_function(lstm_g);
-  InterpreterState lstm_interp(lstm_function);
-  auto outputs = run(lstm_interp, {input[0], hx, cx, w_ih, w_hh});
-  std::tie(hx, cx) = lstm(input[0], hx, cx, w_ih, w_hh);
-
-  // std::cout << almostEqual(outputs[0],hx) << "\n";
-  ASSERT_TRUE(exactlyEqual(outputs[0], hx));
-  ASSERT_TRUE(exactlyEqual(outputs[1], cx));
 }
 
 void testTHNNConv() {
@@ -424,62 +347,11 @@ void testATenNativeBatchNorm() {
   assertAllClose(tensor_grads_out, expected_tensor_grads_out);
 }
 
-using var_meta_type = std::vector<int64_t>;
-using var_meta_list = std::vector<var_meta_type>;
-using test_fn_type = std::function<variable_list(const variable_list&)>;
-
-void testRegisterFusionCachesKernel(std::ostream& out = std::cout) {
-  // Build up a fake graph with a FusionGroup
-  auto createGraphWithNames = [](std::string cname, std::string dname) {
-    auto graph = std::make_shared<Graph>();
-    at::ScalarType s = at::ScalarType::Float;
-    auto type = CompleteTensorType::create(s, at::kCPU, {2, 3, 4}, {12, 4, 1});
-    auto a = SymbolicVariable::asNewInput(*graph, type);
-    auto b = SymbolicVariable::asNewInput(*graph, type);
-    auto c = a * b;
-    auto d = c * a;
-    c.value()->setUniqueName(cname);
-    d.value()->setUniqueName(dname);
-    graph->registerOutput(d.value());
-    torch::jit::overrideCanFuseOnCPU(true);
-    FuseGraph(graph);
-    torch::jit::overrideCanFuseOnCPU(false);
-    return graph;
-  };
-
-  auto getFusionGroup = [](const std::shared_ptr<Graph>& graph) {
-    const auto& nodes = graph->nodes();
-    auto maybe_fusion_group =
-        std::find_if(nodes.begin(), nodes.end(), [](const Node* node) {
-          return node->kind() == prim::FusionGroup;
-        });
-    AT_CHECK(
-        maybe_fusion_group != nodes.end(),
-        "testRegisterFusionCachesKernel: could not create FusionGroup");
-    return *maybe_fusion_group;
-  };
-
-  // Create two alpha-equivalent fusion groups
-  auto graph1 = createGraphWithNames("c1", "d1");
-  auto fg1 = getFusionGroup(graph1);
-
-  auto graph2 = createGraphWithNames("c2", "d2");
-  auto fg2 = getFusionGroup(graph2);
-
-  // Register both with the fusion compiler.
-  auto expected_key = registerFusion(fg1);
-  auto second_key = registerFusion(fg2);
-
-  // Because the graphs are alpha-equivalent, they should return the same key
-  // and therefore share a KernelSpec to share kernels for specializations
-  ASSERT_EQ(second_key, expected_key);
-}
-
 static const auto cf_examples = R"JIT(
   def if_test(a, b):
       # FIXME: use 0 instead of a.
       # c = 0
-      c = a
+        c = a
       if bool(a < b):
         c = b
       else:
@@ -525,205 +397,6 @@ void testControlFlow() {
 void testProto() {
   ::ONNX_NAMESPACE::ModelProto proto;
   proto.set_producer_name("foo");
-}
-
-void testCustomOperators() {
-  {
-    RegisterOperators reg({createOperator(
-        "foo::bar", [](double a, at::Tensor b) { return a + b; })});
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::bar"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::bar");
-
-    ASSERT_EQ(op->schema().arguments().size(), 2);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "_0");
-    ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
-    ASSERT_EQ(op->schema().arguments()[1].name(), "_1");
-    ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
-
-    ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
-
-    Stack stack;
-    push(stack, 2.0f, autograd::make_variable(at::ones(5)));
-    op->getOperation()(stack);
-    at::Tensor output;
-    pop(stack, output);
-
-    ASSERT_TRUE(output.allclose(autograd::make_variable(at::full(5, 3.0f))));
-  }
-  {
-    RegisterOperators reg({createOperator(
-        "foo::bar_with_schema(float a, Tensor b) -> Tensor",
-        [](double a, at::Tensor b) { return a + b; })});
-
-    auto& ops =
-        getAllOperatorsFor(Symbol::fromQualString("foo::bar_with_schema"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::bar_with_schema");
-
-    ASSERT_EQ(op->schema().arguments().size(), 2);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "a");
-    ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
-    ASSERT_EQ(op->schema().arguments()[1].name(), "b");
-    ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
-
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
-
-    Stack stack;
-    push(stack, 2.0f, autograd::make_variable(at::ones(5)));
-    op->getOperation()(stack);
-    at::Tensor output;
-    pop(stack, output);
-
-    ASSERT_TRUE(output.allclose(autograd::make_variable(at::full(5, 3.0f))));
-  }
-  {
-    // Check that lists work well.
-    RegisterOperators reg({createOperator(
-        "foo::lists(int[] ints, float[] floats, Tensor[] tensors) -> float[]",
-        [](const std::vector<int64_t>& ints,
-           const std::vector<double>& floats,
-           std::vector<at::Tensor> tensors) { return floats; })});
-
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::lists"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::lists");
-
-    ASSERT_EQ(op->schema().arguments().size(), 3);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "ints");
-    ASSERT_TRUE(
-        op->schema().arguments()[0].type()->isSubtypeOf(ListType::ofInts()));
-    ASSERT_EQ(op->schema().arguments()[1].name(), "floats");
-    ASSERT_TRUE(
-        op->schema().arguments()[1].type()->isSubtypeOf(ListType::ofFloats()));
-    ASSERT_EQ(op->schema().arguments()[2].name(), "tensors");
-    ASSERT_TRUE(
-        op->schema().arguments()[2].type()->isSubtypeOf(ListType::ofTensors()));
-
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_TRUE(
-        op->schema().returns()[0].type()->isSubtypeOf(ListType::ofFloats()));
-
-    Stack stack;
-    push(stack, std::vector<int64_t>{1, 2});
-    push(stack, std::vector<double>{1.0, 2.0});
-    push(stack, std::vector<at::Tensor>{autograd::make_variable(at::ones(5))});
-    op->getOperation()(stack);
-    std::vector<double> output;
-    pop(stack, output);
-
-    ASSERT_EQ(output.size(), 2);
-    ASSERT_EQ(output[0], 1.0);
-    ASSERT_EQ(output[1], 2.0);
-  }
-  {
-    RegisterOperators reg(
-        "foo::lists2(Tensor[] tensors) -> Tensor[]",
-        [](std::vector<at::Tensor> tensors) { return tensors; });
-
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::lists2"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::lists2");
-
-    ASSERT_EQ(op->schema().arguments().size(), 1);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "tensors");
-    ASSERT_TRUE(
-        op->schema().arguments()[0].type()->isSubtypeOf(ListType::ofTensors()));
-
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_TRUE(
-        op->schema().returns()[0].type()->isSubtypeOf(ListType::ofTensors()));
-
-    Stack stack;
-    push(stack, std::vector<at::Tensor>{autograd::make_variable(at::ones(5))});
-    op->getOperation()(stack);
-    std::vector<at::Tensor> output;
-    pop(stack, output);
-
-    ASSERT_EQ(output.size(), 1);
-    ASSERT_TRUE(output[0].allclose(autograd::make_variable(at::ones(5))));
-  }
-  {
-    auto op = createOperator(
-        "traced::op(float a, Tensor b) -> Tensor",
-        [](double a, at::Tensor b) { return a + b; });
-
-    std::shared_ptr<tracer::TracingState> state;
-    std::tie(state, std::ignore) = tracer::enter({});
-
-    Stack stack;
-    push(stack, 2.0f, autograd::make_variable(at::ones(5)));
-    op.getOperation()(stack);
-    at::Tensor output = autograd::make_variable(at::empty({}));
-    pop(stack, output);
-
-    tracer::exit({IValue(output)});
-
-    std::string op_name("traced::op");
-    bool contains_traced_op = false;
-    for (const auto& node : state->graph->nodes()) {
-      if (std::string(node->kind().toQualString()) == op_name) {
-        contains_traced_op = true;
-        break;
-      }
-    }
-    ASSERT_TRUE(contains_traced_op);
-  }
-  {
-    ASSERT_THROWS_WITH(
-        createOperator(
-            "foo::bar_with_bad_schema(Tensor a) -> Tensor",
-            [](double a, at::Tensor b) { return a + b; }),
-        "Inferred 2 argument(s) for operator implementation, "
-        "but the provided schema specified 1 argument(s).");
-    ASSERT_THROWS_WITH(
-        createOperator(
-            "foo::bar_with_bad_schema(Tensor a) -> Tensor",
-            [](double a) { return a; }),
-        "Inferred type for argument #0 was float, "
-        "but the provided schema specified type Tensor "
-        "for the argument in that position");
-    ASSERT_THROWS_WITH(
-        createOperator(
-            "foo::bar_with_bad_schema(float a) -> (float, float)",
-            [](double a) { return a; }),
-        "Inferred 1 return value(s) for operator implementation, "
-        "but the provided schema specified 2 return value(s).");
-    ASSERT_THROWS_WITH(
-        createOperator(
-            "foo::bar_with_bad_schema(float a) -> Tensor",
-            [](double a) { return a; }),
-        "Inferred type for return value #0 was float, "
-        "but the provided schema specified type Tensor "
-        "for the return value in that position");
-  }
-  {
-    // vector<double> is not supported yet.
-    auto op = createOperator(
-        "traced::op(float[] f) -> int",
-        [](const std::vector<double>& f) -> int64_t { return f.size(); });
-
-    std::shared_ptr<tracer::TracingState> state;
-    std::tie(state, std::ignore) = tracer::enter({});
-
-    Stack stack;
-    push(stack, std::vector<double>{1.0});
-
-    ASSERT_THROWS_WITH(
-        op.getOperation()(stack),
-        "Tracing float lists currently not supported!");
-
-    tracer::abandon();
-  }
 }
 
 void testEvalModeForLoadedModule() {
@@ -900,196 +573,6 @@ void testTopologicalIndex() {
       }
     }
   }
-}
-
-std::unique_ptr<detail::DynamicDAG<std::string>> newDynamicDAG() {
-  return std::unique_ptr<detail::DynamicDAG<std::string>>(
-      new detail::DynamicDAG<std::string>());
-}
-
-void testNewVertex() {
-  auto graph = newDynamicDAG();
-  AT_ASSERT(graph->debugNumVertices() == 0);
-  auto a = graph->newVertex("a");
-  AT_ASSERT(graph->debugNumVertices() == 1);
-  AT_ASSERT(a->ord == 0);
-  AT_ASSERT(a->data.size() == 1);
-  AT_ASSERT(a->data[0] == "a");
-  AT_ASSERT(a->in_edges().size() == 0);
-  AT_ASSERT(a->out_edges().size() == 0);
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  AT_ASSERT(graph->debugNumVertices() == 3);
-  AT_ASSERT(b->ord == 1);
-  AT_ASSERT(c->ord == 2);
-}
-
-void testAddEdgeBasic() {
-  // a -> b -> c
-  // \---------^
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  graph->addEdge(a, b);
-  graph->addEdge(b, c);
-  graph->addEdge(a, c);
-  AT_ASSERT(a->in_edges().size() == 0);
-  AT_ASSERT(a->out_edges().size() == 2);
-  AT_ASSERT(a->out_edges().contains(b));
-  AT_ASSERT(a->out_edges().contains(c));
-  AT_ASSERT(b->in_edges().size() == 1);
-  AT_ASSERT(b->out_edges().size() == 1);
-  AT_ASSERT(b->in_edges().contains(a));
-  AT_ASSERT(b->out_edges().contains(c));
-  AT_ASSERT(c->in_edges().size() == 2);
-  AT_ASSERT(c->out_edges().size() == 0);
-  AT_ASSERT(c->in_edges().contains(a));
-  AT_ASSERT(c->in_edges().contains(b));
-}
-
-void testAddEdgeCycleDetection() {
-  // a -> b -> c
-  // ^---------/
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  graph->addEdge(a, b);
-  graph->addEdge(b, c);
-  bool erred = false;
-  try {
-    graph->addEdge(c, a);
-  } catch (c10::Error& err) {
-    erred = true;
-  }
-  AT_ASSERT(erred);
-}
-
-void testAddEdgeReordersBasic() {
-  // a, b => b -> a
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  AT_ASSERT(a->ord == 0);
-  AT_ASSERT(b->ord == 1);
-  graph->addEdge(b, a);
-  AT_ASSERT(a->ord == 1);
-  AT_ASSERT(b->ord == 0);
-}
-
-void testAddEdgeReordersComplicated() {
-  // a -> b  c -> d with addEdge(d, b) ==>
-  // c -> d -> a -> b
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  auto d = graph->newVertex("d");
-  graph->addEdge(a, b);
-  graph->addEdge(c, d);
-  AT_ASSERT(a->ord == 0);
-  AT_ASSERT(b->ord == 1);
-  AT_ASSERT(c->ord == 2);
-  AT_ASSERT(d->ord == 3);
-  graph->addEdge(d, a);
-  AT_ASSERT(c->ord == 0);
-  AT_ASSERT(d->ord == 1);
-  AT_ASSERT(a->ord == 2);
-  AT_ASSERT(b->ord == 3);
-  AT_ASSERT(c->in_edges().size() == 0);
-  AT_ASSERT(c->out_edges().size() == 1);
-  AT_ASSERT(c->out_edges().contains(d));
-  AT_ASSERT(d->in_edges().size() == 1);
-  AT_ASSERT(d->out_edges().size() == 1);
-  AT_ASSERT(d->in_edges().contains(c));
-  AT_ASSERT(d->out_edges().contains(a));
-  AT_ASSERT(a->in_edges().size() == 1);
-  AT_ASSERT(a->out_edges().size() == 1);
-  AT_ASSERT(a->in_edges().contains(d));
-  AT_ASSERT(a->out_edges().contains(b));
-  AT_ASSERT(b->in_edges().size() == 1);
-  AT_ASSERT(b->out_edges().size() == 0);
-  AT_ASSERT(b->in_edges().contains(a));
-}
-
-void testRemoveEdgeBasic() {
-  // a -> b
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  graph->addEdge(a, b);
-  AT_ASSERT(graph->debugNumVertices() == 2);
-  graph->removeEdge(a, b);
-  AT_ASSERT(graph->debugNumVertices() == 2);
-  AT_ASSERT(a->out_edges().size() == 0);
-  AT_ASSERT(b->in_edges().size() == 0);
-}
-
-void testRemoveVertexBasic() {
-  // a -> b
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  graph->addEdge(a, b);
-  graph->addEdge(b, c);
-  AT_ASSERT(graph->debugNumVertices() == 3);
-  graph->removeVertex(b);
-  AT_ASSERT(graph->debugNumVertices() == 2);
-  AT_ASSERT(a->out_edges().size() == 0);
-  AT_ASSERT(c->in_edges().size() == 0);
-}
-
-void testContractEdgeBasic() {
-  // a -> b -> c -> d
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  auto d = graph->newVertex("d");
-  graph->addEdge(a, b);
-  graph->addEdge(b, c);
-  graph->addEdge(c, d);
-  graph->contractEdge(b, c);
-  AT_ASSERT(graph->debugNumVertices() == 3);
-  AT_ASSERT(a->out_edges().size() == 1);
-  AT_ASSERT(d->in_edges().size() == 1);
-  AT_ASSERT(*a->out_edges().begin() == *d->in_edges().begin());
-  auto* contracted = *a->out_edges().begin();
-  AT_ASSERT(contracted->data.size() == 2);
-  AT_ASSERT(contracted->data[0] == "b");
-  AT_ASSERT(contracted->data[1] == "c");
-  AT_ASSERT(contracted->out_edges().size() == 1);
-  AT_ASSERT(contracted->in_edges().size() == 1);
-  AT_ASSERT(contracted->in_edges().contains(a));
-  AT_ASSERT(contracted->out_edges().contains(d));
-}
-
-void testContractEdgeCycleDetection() {
-  // a -> b -> c
-  // `---------^
-  // contractEdge(a, c) will cause a cycle
-  auto graph = newDynamicDAG();
-  auto a = graph->newVertex("a");
-  auto b = graph->newVertex("b");
-  auto c = graph->newVertex("c");
-  graph->addEdge(a, b);
-  graph->addEdge(b, c);
-  graph->addEdge(a, c);
-  AT_ASSERT(!graph->contractEdge(a, c));
-}
-
-void testDynamicDAG() {
-  testNewVertex();
-  testAddEdgeBasic();
-  testAddEdgeCycleDetection();
-  testAddEdgeReordersBasic();
-  testAddEdgeReordersComplicated();
-  testRemoveEdgeBasic();
-  testRemoveVertexBasic();
-  testContractEdgeBasic();
-  testContractEdgeCycleDetection();
 }
 
 void testAutogradProfiler() {
